acv+rbo: COR-1749
- Introduced NetworkingMBean and tests (java only / bytes sent only)
- Socket4#write(int) => Socket4#write(byte)
- Introduced Socket4Decorator
- Introduced ServerSocket4Decorator
- Run methods within db4o environment where appropriated
- Optional project depends on CS

+++ db4o.cs/src/com/db4o/foundation/network/Socket4.java
@@ -20,7 +20,7 @@ public interface Socket4 {
 
   	void write(byte[] bytes, int offset, int count) throws IOException;
   	
-  	void write(int b) throws IOException;
+  	void write(byte b) throws IOException;
 
 	Socket4 openParalellSocket() throws IOException;
